requirements.txt
streamlit==1.28.0
tensorflow==2.13.0
keras==2.13.1
numpy==1.26.2
pandas==2.1.0
Pillow==10.1.0
scikit-learn==1.3.2
matplotlib==3.8.0
